requirements.txt
PyYaml
appdirs
mailjet_rest
click
sphinx
